libdfp: Enable BID for 32bits
This patch adds BID support for 32 bits. BID current relies on 128-bit
type support, which is supplied by GCC/clang extension __uint128_t.
Some architectures, however, does not support this extension
(x86 32 bits, for instance).

In these case the mini-gmp implementation from libgmp is used to supply
a multiprecision math library (set to emulate a 128-bit int type).
Currently it is included direct in libdecnumber/decimal128.c object and
thus used in other parts of libdfp through libdecnumber.a static
library.

+++ sysdeps/bid/bid-private.c
@@ -103,7 +103,7 @@ void
 __get_digits_d64 (_Decimal64 x, char *str, int *exp_p, int *sign_p,
 		  int *nan_p, int *inf_p)
 {
-  unsigned long int result;
+  unsigned long long int result;
   int i, size, exp = 0;
 
   struct ieee754r_c_field c_f;
@@ -138,11 +138,11 @@ __get_digits_d64 (_Decimal64 x, char *str, int *exp_p, int *sign_p,
 
   exp -= DECIMAL64_Bias;
 
-  sprintf (str, "%lu", result);
+  sprintf (str, "%llu", result);
   size = (16 - strlen (str));
   for (i = 0; i < size; i++)
     str[i] = '0';
-  sprintf (str + i, "%lu", result);
+  sprintf (str + i, "%llu", result);
   str[16] = '\0';
 
   if (sign_p)
@@ -164,7 +164,7 @@ __get_digits_d128 (_Decimal128 x, char *str, int *exp_p, int *sign_p,
 
   struct ieee754r_c_field c_f;
   union ieee754r_Decimal128 d;
-  unsigned long int result[2];
+  unsigned long long int result[2];
 
   d.td = x;
   c_f = c_decoder[d.nan.nan];
